+++ src/light/mod.rs
@@ -10,5 +10,5 @@ pub trait Light: std::fmt::Debug {
 /// Represent a light which has an abstract position in the scene being rendered.
 pub trait SpatialLight: Light {
     /// Get a unit vector from the origin to the position of the light, and its distance
-    fn to_source(&self, origin: &Point) -> (Vector, f64);
+    fn to_source(&self, origin: &Point) -> (Vector, f32);
 }
